Get data and display markers in the UI
This commit sets up an AJAX call to the server. When the server has responded, a new Google Maps Map instance is instantiated and placed in the UI. Each datapoint in the response body is translated into a Marker on the map with the species common name as the marker title.