Make README properly display datatypes in the JSON
This commit fixes a markdown formatting issue where the datatypes in angle brackets in the JSON were being interpreted as empty HTML tags and weren't being displayed on the GitHub repository page. Now, the section of the README detailing the JSON format is being interpreted as code instead of as markdown, which makes the datatypes in angle brackets be displayed as intended.