Clean up more formatting in README
This commit finishes what the previous commit started. It fixes the rest of the places where code in angle brackets was being interpreted as empty HTML tags and was not being displayed in the README on the GitHub repository page.

As of now, all of the code snippets in angle brackets are displayed as intended.